Fixed infinite loops from weird imports.
