+++ preprocess.py
@@ -7,15 +7,10 @@ import sys
 
 from preprocess_setup_launch import simplify_image, simplify_mask
 
-
-def batch_process():
+if __name__ == "__main__":
 	f = open(sys.argv[1])  # This is the name of the file containing timestamps
 	print("Opened {}".format(sys.argv[1]))
 	for time in f:
 		time = time.replace('\n', '')
 		simplify_mask(time)
 		simplify_image(time)
-
-
-if __name__ == "__main__":
-	batch_process()
